+++ NEWS.md
@@ -5,10 +5,11 @@
 - fix: CompressedBand now always returns array when indexed with a slice
 - fix: output dimension of grid sampling methods is now always one larger than
   the input
+- `merge` now works with multiband grids
+- use more efficient algorithm for intersection detection
 - new `read_gtiffs` function for reading multiple files as bands
 - implement `RegularGrid.coordinates()` method returning a `CoordGenerator` object
 - implement `get_coordinate_lists` for Multiline and Multipolygon
-- `merge` now works with multiband grids
 - performance optimizations and fixes
 
 ## changes with 0.7.1